Add reservation ID to record type
There's essentially two ways to handle entity IDs in CRUD scenarios:
either you supply an entity without ID and generate it on the server, or
else you let it be the client's responsibility to supply a unique ID.

In the first case, you'll need two record types: one to receive a CREATE
request, and one for all other cases. The first of these types have no
ID field, but the other one has. In all other respects, the two types
would be identical.

In the second case, you save one of those two types, and make the
internal API more consistent, but the disadvantage (if it's one) is that
the client must supply a unique ID. This can be addressed by using GUIDs
as IDs.

In this code base I choose to go with the simpler option, as I imagine
that all clients will be other programs, which are perfectly capable of
generating GUIDs as required.