Introduce free ReservationsProgram monad
At the moment, this doesn't add more value, but is the first step
towards being able to implement a more complex function to handle POST
requests in a testable manner. This commit just sets up some of the
infrastructure.

The reservationServer now receives an interpreter as a function
argument. Later, the read-side of the API is going to need some
implementation code as well, so this could be an opportunity to show how
a DI Container in FP can be made with the Reader monad and some
contravariant mapping...