Change reservationDate to LocalTime
While a reservation time can be viewed as an instant in time, what
really matters is that it's a local time. Recording times local to the
restaurant in question is better aligned with what the restaurant and the
customer actually needs to know.

If recording a reservantion time in ZonedTime, you run the risk that
when you book sufficiently long in advance, you may not know exactly
what time zone offset will be in effect at that date. These are, after
all, politically determined, and may not be entirely predictable.