Add argument to ReadReservations instruction
The intent of this operation is to find reservations in a particular
range - typically on the same day, but beginning at different times. The
logic to find the low and high value should be part of the business
logic so that it can be unit tested. The database query should be as
simple and dumb as possible.