Add test of 404 reservation
This, honestly, required the ceremonial shaving of several yaks. I tried
to make the test server run in StateT, but it seems that it doesn't
bring the state forward between 'HTTP' requests. At least, I couldn't
get it to work, but with IORef, it does work.

While I consider it fine and proper to express integration tests
entirely through the public API, the apparent lack of backdoor
manipulation worries me a bit.