Remove Validation module
It turns out that Servant's built-in validation (for the Accept
instances and associated MimeRender and MimeUnrender instances, IIRC...)
uses Either (or, rather, ExceptT) to return HTTP errors, so applicative
validation wouldn't quite fit that mold.

It could be made to work, but it'd only work for the data components
that are explicitly validated. It would have no effect for those that
are already handled by Servant, which only returns an error message for
the first problem it encounters.