Unit test validation of 'less than'
This may seem like a bit of a redundant set of unit tests, but when I
originally implemented the feature that checks the date of a
reservation, I managed to get confused and return `x` from the function,
instead of `y`.

My philosophy is that if someone manages to make a mistake in the
implementation of a function, then that function is complicated enough
to warrant a test - at least a test that reproduces the original
mistake. That's what the first of these two tests does.

The second test, then, makes sure that both branches are covered by
automated tests.

I've seen both tests fail by changing the definition of
`validateLessThan`.