Introduce APIError type
The pattern match made in API.reservationServer is currently incorrect,
but necessary to avoid a compiler warning. A future test should flush
out the appropriate behaviour.

The purpose of the new type is to be able to distinguish between
validation/input errors (400 HTTP error codes) and genuine execution
errors (500 HTTP error codes).