Validate reservation ID
The `suchThat` filter on `ValidReservation` in the test doesn't seem to
be strictly necessary, as the tests also pass without it. I suppose that
the UUID Arbitrary instance doesn't explicitly generate that well-known
value, although I think that it ought to. For good measure, however, I
added the filter to the `ValidReservation` `Arbitrary` instance.