Rename to canAccommodate
This is the first of a few refactorings that'll turn this function into
a function that specifically looks for a free table with sufficient
capacity.

While I originally thought I'd handle the requirements differently, I'm
now pursuing a strategy where the canAccommodate function will look for
a free table, while a preliminary filter will remove all those
reservations that aren't relevant because they don't overlap in time.