Make canAccommodate generic
This may turn out to be too generic in terms of what's actually
required, but I don't mind having to narrow the choice of containers, as
long as I can keep the values generic.

The motivation for this move is that the more generic the function
becomes, the less it can do, and thus, the fewer tests are required. See
e.g. https://blog.ploeh.dk/2018/07/09/typing-and-testing-problem-23 for
an exploration of parametric polymorphism and free theorems.