Query database for existing reservations in range
There's still bugs in the implementation. One is that the query for
existing databases only queries for reservations for the current time
and going forward. This means that you can have seatings in progress at
the time of the reservation that aren't detected. An additional test
ought to reproduce that bug.