Add explicit `runSessionWithApp` function
It dawned on me that the current `withApp` function specifically
produces a `Property`, which is fine for testing a property with
QuickCheck, but isn't going to help if I want to write an HUnit test
that exercises the application. Since I haven't done so yet, I hadn't
noticed this limitation.