Add sample connection string
I finally got around to figure out how to connect to a local SQL Server
Express instance. This sample connection string shows how to connect to
an already existing database instance called 'Booking'.

Before that, I'd been using a SQL Azure instance. One can find ODBC
connection strings for one's databases in the Azure portal, but in order
for this to work, one has to replace the driver version from the
suggested 13 to 17, as also shown in this file.

In both cases, the database should contain the Reservations table
defined in BookingDbSchema.sql.