Add .vscode to .gitignore
It seems to be a newer thing that Visual Studio code adds this .vscode
directory. It's not clear whether one is supposed to commit the files in
this directory, but at the moment, it contains no data anyway, so it
seems reasonable to just ignore it.

I also added .vs because I sometimes open the files in Visual Studio(!).